MJPEG streaming done and worked on vitals

+++ tmobile-5g-hackathon/Models/NetworkModel.swift
@@ -26,11 +26,11 @@ struct GET_ambulances_Response_item: Decodable {
 // WS /ambulance_data?ambulance_id=3342342342342423 VITALS_UPDATE
 struct WS_VITALS_UPDATE: Decodable {
     var time_ms: Int64
-    var ekg_value: Float?
-    var heart_rate: Int?
-    var blood_oxygen_percent: Int?
-    var diastolic_blood_pressure: Int?
-    var systolic_blood_pressure: Int?
+    var ekg_value: Double?
+    var heart_rate: Double?
+    var blood_oxygen_percent: Double?
+    var diastolic_blood_pressure: Double?
+    var systolic_blood_pressure: Double?
 }
 
 // WS /ambulance_data?ambulance_id=3342342342342423 NEW_FILE
